Maps now works, gets correct LatLng from array, Zooms automatically but no control over it, need to fix it + fix no buttons on map. Maps can open in a new page automatically (starts Maps Application) Map UI unusable due to scrollview. Still need to fix
Need to add photos to the gallery
Still need to redesign and make it a bit better

+++ app/src/main/java/uom/ict/mdp/Events.java
@@ -139,11 +139,11 @@ public class Events implements Comparable<Events> {
         this.location = location;
     }
 
-    public double getxCoordinates() {return xCoordinates;}
+    public float getxCoordinates() {return xCoordinates;}
 
     public void setxCoordinates(float xCoordinates) {this.xCoordinates = xCoordinates;}
 
-    public double  getyCoordinates() {return yCoordinates;}
+    public float  getyCoordinates() {return yCoordinates;}
 
     public void setyCoordinates(float yCoordinates) {this.yCoordinates = yCoordinates;}
